Refactor generator.go and prepare for datasource integration
The code in generator.go has been modularized into more specific functions for creating a file path and writing to a file. These modifications enhance the readability of the code and conform with the single-responsibility principle. Comment placeholders were also added in codegen.go to earmark places for future integration with datasources. Moreover, the fixed class name 'ExampleResource' was replaced with a dynamic one in terraform_resource.go.