Updated LCS implementation and added OptimizeMovements step
Fixed some bugs in LCS implementation, but it needs to be be optimized - it takes ~25 seconds to find a common
subsequence in two equal sequences of 50k elements.

Added an OptimizeMovements step that is done after GenerateMovements returns a list of all move actions. It removes
redundant actions that, after previous actions were applied, no longer have any effect.