Add imports location lock to serialize non-atomic operations
Adding and removing items from the imports location list is currently
implemented in a non-atomic way, by first reading a list of items from the
server, updating it locally, and sending a new list to the server.

Introduce a (per-resource, per-location) lock to all resources that require
modifications of the import list so that those operations are done sequentially.