fix(terraform): Handle position attribute within lifecycle.ignore_changes
When position is included in lifecycle.ignore_changes, it's possible for position to become
null on refresh - this is due to the fact that Read() marks position as out of sync by setting
it to null value. Normally, this triggers a plan change which is then applied, but if position
is listed in ignore_changes terraform doesn't generate a plan for it, and leaves position in state
as null.